requirements.txt
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
docopt==0.6.2
idna==3.3
Js2Py==0.71
packaging==21.3
pipwin==0.5.1
PyAudio==0.2.11
pycodestyle==2.8.0
pyjsparser==2.7.1
pyparsing==3.0.6
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==0.19.2
pyttsx3==2.90
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wikipedia==1.4.0
